requirements.txt
asgiref==3.8.1
attrs==24.2.0
cfgv==3.4.0
coverage==7.6.1
distlib==0.3.8
Django==5.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
filelock==3.15.4
identify==2.6.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg==3.2.1
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.26.3
